refactor(namedquery): extract Flight-free parts to commons
Move the reusable named-query types out of the Flight module so non-Flight
consumers can reuse them without depending on dazzleduck-sql-flight:

- commons.namedquery: NamedQueryDefinition/Request/Response (moved), plus new
  NamedQueryStore (DuckDB template load via ConnectionPool) and
  NamedQueryValidators (validate + describe validator class names).
- DefaultNamedQueryServiceAdaptor reduced to orchestration: store.loadByName
  -> validators.validate -> Jinjava render -> Flight execute.

Records require Java 16+, so these live in commons (Java 21), not common
(Java 11). Jinjava rendering stays in flight by design.
